// frontend/src/types/HybridRecommender.ts
export interface Recommend {
    originalMovieId?: string;
    originalMovieTitle?: string;
    rec1Id?: string;
    rec1Title?: string;
    rec1Score?: number;
    rec2Id?: string;
    rec2Title?: string;
    rec2Score?: number;
    rec3Id?: string;
    rec3Title?: string;
    rec3Score?: number;
    rec4Id?: string;
    rec4Title?: string;
    rec4Score?: number;
    rec5Id?: string;
    rec5Title?: string;
    rec5Score?: number;
}